% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{mw_filter}
\alias{mw_filter}
\title{Filter healthcare wait time data}
\usage{
mw_filter(
  df,
  province = NULL,
  indicator = NULL,
  start_year = NULL,
  end_year = NULL
)
}
\arguments{
\item{df}{A standardized data frame}

\item{province}{Province name(s) to filter}

\item{indicator}{Keyword used to match indicators}

\item{start_year}{Starting year for filtering}

\item{end_year}{Ending year for filtering}
}
\value{
A filtered data frame
}
\description{
Filter healthcare wait time data
}
