% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_download.R
\name{mw_list_downloads}
\alias{mw_list_downloads}
\title{List downloaded CSV files in cache}
\usage{
mw_list_downloads(cache_dir = NULL)
}
\arguments{
\item{cache_dir}{cache directory}
}
\value{
tibble of file paths
}
\description{
List downloaded CSV files in cache
}
