% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_visualize.R
\name{mw_plot_affected_province}
\alias{mw_plot_affected_province}
\title{Bar plot: provinces affected by wait time}
\usage{
mw_plot_affected_province(std, year = 2024)
}
\description{
Bar plot: provinces affected by wait time
}
