% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_visualize.R
\name{mw_plot_compare}
\alias{mw_plot_compare}
\title{Compare values across groups for a single year (bar chart)}
\usage{
mw_plot_compare(df, year, group_by = "geo", title = NULL)
}
\arguments{
\item{df}{Standardized/filtered data (output of mw_standardize() or mw_filter())}

\item{year}{Single year to compare (e.g., 2022)}

\item{group_by}{Column to compare across (default "geo" or you can use "indicator")}

\item{title}{Optional plot title}
}
\value{
A ggplot object
}
\description{
Compare values across groups for a single year (bar chart)
}
