% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_visualize.R
\name{mw_plot_satisfaction_stack}
\alias{mw_plot_satisfaction_stack}
\title{Stacked (100\%) bar: satisfaction vs dissatisfaction by province
Uses "Number of persons" to compute within-province percentages}
\usage{
mw_plot_satisfaction_stack(std, year = 2024)
}
\description{
Stacked (100\%) bar: satisfaction vs dissatisfaction by province
Uses "Number of persons" to compute within-province percentages
}
