% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_visualize.R
\name{mw_plot_wait_duration_stack}
\alias{mw_plot_wait_duration_stack}
\title{Stacked (100\%) bar: wait duration distribution by province
Uses "Number of persons" to compute within-province percentages}
\usage{
mw_plot_wait_duration_stack(std, year = 2024)
}
\arguments{
\item{std}{Standardized medwait data (output of mw_standardize() and/or mw_filter()).
Must include at least: ref_date, stat, indicator, geo, value.}

\item{year}{Integer year to plot (default 2024).}
}
\value{
A ggplot object
}
\description{
Stacked (100\%) bar: wait duration distribution by province
Uses "Number of persons" to compute within-province percentages
}
