% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_retrieve.R
\name{mw_read_data}
\alias{mw_read_data}
\title{Read a downloaded main CSV into tibble (minimal wrangling)}
\usage{
mw_read_data(product_id, language = "en")
}
\arguments{
\item{product_id}{StatCan product ID}

\item{language}{"en" or "fr"}
}
\value{
tibble
}
\description{
Read a downloaded main CSV into tibble (minimal wrangling)
}
