% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_retrieve.R
\name{mw_read_metadata}
\alias{mw_read_metadata}
\title{Read a downloaded MetaData CSV into tibble}
\usage{
mw_read_metadata(product_id, language = "en")
}
\arguments{
\item{product_id}{StatCan product ID}

\item{language}{"en" or "fr"}
}
\value{
tibble
}
\description{
Read a downloaded MetaData CSV into tibble
}
