% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c_filter.R
\name{mw_filter}
\alias{mw_filter}
\title{Filter standardized plot-ready data}
\usage{
mw_filter(
  df,
  geo = NULL,
  indicator = NULL,
  stat = NULL,
  start_year = NULL,
  end_year = NULL,
  exact_indicator = FALSE
)
}
\arguments{
\item{df}{Standardized tibble (output of mw_standardize)}

\item{geo}{GEO values to keep (character vector)}

\item{indicator}{Keyword(s) or exact labels (character vector)}

\item{stat}{Statistic/characteristic filter (e.g., "Percent", "Number of persons")}

\item{start_year}{Inclusive start year}

\item{end_year}{Inclusive end year}

\item{exact_indicator}{If TRUE, exact match; else keyword match}
}
\value{
Filtered tibble
}
\description{
Filter standardized plot-ready data
}
